refactor: extract REVEAL auto-advance cluster from state.py (#1271)
Pull the #1012 unattended REVEAL→next-round machinery out of the
state.py God-Object into a new RevealAutoAdvanceMixin
(game/state_auto_advance.py). Pure code-move, byte-identical method
bodies — no behavior change.

Moved:
* _cancel_auto_advance — supersede the pending auto-advance task
* _song_finished       — song-end poll helper (media-player playback state)
* _reveal_auto_advance — song-end / dwell auto-advance task
* _reveal_idle_halt    — zero-guesses hold-on-REVEAL task

Stays on GameState (SSOT / round-end orchestration):
* _schedule_reveal_advance (the scheduler that starts these tasks)
* _set_phase chokepoint, _score_all_players, _end_round_unlocked

Mixin added alphabetically to the GameState base list + a class-docstring
paragraph. GamePhase imported lazily inside the methods (# noqa: PLC0415)
to avoid the state.py circular import, matching the existing state_*.py
mixins.

state.py: 1469 -> 1360 lines (-109), progressing #1271 toward <800.

Gates (all green, no annotation fixes needed):
* pytest: 885 passed, 2 skipped
* ruff check: passed
* ruff format --check: passed
* mypy 1.18.2: Success, no issues in 15 source files
